finished basic form framework, need to work through props and state
