+++ frontend/src/Components/Cards.jsx
@@ -1,0 +1,26 @@
+import React from 'react';
+import {
+  Card, CardText, CardBody, CardLink,
+  CardTitle, CardSubtitle
+} from 'reactstrap';
+
+const Cards = (props) => {
+  return (
+    <div>
+      <Card>
+        <CardBody>
+          <CardTitle tag="h5">Card title</CardTitle>
+          <CardSubtitle tag="h6" className="mb-2 text-muted">Card subtitle</CardSubtitle>
+        </CardBody>
+        <img width="100%" src="/assets/318x180.svg" alt="Card image cap" />
+        <CardBody>
+          <CardText>Some quick example text to build on the card title and make up the bulk of the card's content.</CardText>
+          <CardLink href="#">Card Link</CardLink>
+          <CardLink href="#">Another Link</CardLink>
+        </CardBody>
+      </Card>
+    </div>
+  );
+};
+
+export default Cards;
