+++ frontend/src/Components/Navigation.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Navigation() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Navigation
